Add files via upload
This code is a simple number guessing game implemented in Python. The game allows the user to select a range of integers from A to B, where both A and B are integers provided by the user. The system then randomly selects an integer within this range, and the user's objective is to guess this integer in the fewest number of attempts possible.